move draw_bb
